add support for raw TCP socket upgrading
This is supported by the /exec/(id)/start and /containers/(id)/attach
endpoints, allowing the HTTP request/response to be hijacked and turned
in to a regular TCP socket. This allows for useful features such as
half-closed sockets, which properly represents how a command's stdin
can be closed and it can still write output to stdout.

Requires HTTP->TCP connection 'upgrading' from docker-modem, which is
proposed in apocas/docker-modem#54.